Remove the helper and the need to double include
This is a simpler way to do this
